Moved CheckboxState enum in (new) Enum.hpp
